requirements.txt
beautifulsoup4==4.10.0
bs4==0.0.1
docopt==0.6.2
dnspython==2.1.0
geoip2==4.1.0
ipython==7.30.1
jupyter-client==7.1.0
jupyter-core==4.9.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyshodan==3.0.0
requests==2.26.0
shodan==2.0.0
bitcoin
eth-account==0.8.0
dnspython==2.3.0
ipaddress
pytest
exifread==3.0.0